+++ truehd/src/utils/mod.rs
@@ -8,4 +8,5 @@ pub mod buffer_pool;
 pub mod crc;
 pub mod dither;
 pub mod errors;
+pub mod perf;
 pub mod timing;